added possibility to pass own list of expressions. made regex quantifiers lazy.
the path to an existing txt file containing expressions can be passed to the script, making it use the defined expression-set. in addition to this the regex quantifiers were made lazy, to prevent them selecting the whole line following. this could happen, because sometimes the closing curly bracket would not be the last in line. by defining an indefinite amount of characters between the starting and closing curly bracket, whole lines were deleted.

Signed-off-by: Only-a-User <49812354+Only-a-User@users.noreply.github.com>